[LCR-238] refactor: 오늘 입장자 수 조회 기능 테스트 코드 리팩토링
